Initial Agent commit
Adds structure for the distributed erlang (YOLO) application.
Fleshes out the first revision of the network mediation process
(responsible for handling requests and inter/intra-agent communication).

It is now possible to build a graph of netmed processes, send one a
message, and watch the information propagate through the network.

Local services are treated the same as remote ones, but are given the
special atom 'local'.